tests/wpt/tests/input-events/contenteditable-insertfrompaste-type-inputevent-data.html

<!DOCTYPE html>
<html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="../editing/include/editor-test-utils.js"></script>
<div contenteditable id="editable">Copying and pasting some part of the text should not
  set the event.data with the selected part for contenteditable</div>
<script>
  const editable = document.getElementById("editable");
  let eventData;
  let eventFired = false;
  editable.addEventListener("input", (evt) => {
    if(evt.inputType == "insertFromPaste") {
      eventFired = true;
      eventData = evt.data;
    }
  });
  promise_test(async () => {
    editable.focus();
    // Selecting first word of the text.
    getSelection().setBaseAndExtent(editable.firstChild, 0, editable.firstChild, 7);
    const selectedData = getSelection().toString();
    assert_equals(selectedData, "Copying");
    // Copy and paste should fire input event with inputType insertFromPaste.
    const utils = new EditorTestUtils(editable);
    await utils.sendCopyShortcutKey();
    await utils.sendPasteShortcutKey();
    // Event data should NOT be set with the first word of the text
    // which is "Copying" in case of contenteditable, it should be null instead.
    assert_equals(eventData, null);
    assert_true(eventFired);
  }, "Input event data for inputType insertFromPaste should not be set for contenteditable");
</script>
</body>
</html>
